enable chunking of AppendEntries
gRPC has a limit on message size it accepts to receive. By default, in
the Go implementation this limit is 4MB. If a raft log entry if bigger
than 4MB (or the configured limit on the remote), the AppendEntries call
will fail. Equally, an AppendEntries can contains up to 64 (by default)
log entries. If the sum of all log entries is bigger than the message
limit, the call is reject. The same limit applies to
AppendEntriesPipeline.

This commit add support to "chunk" the AppendEntries call so that if it
exceeds the maximum size, it can still be transfered in multiple chunks.
It is backward compatible because it fallback to the previous behavior
if the remote does not support chunking because its raft-grpc-transport
is not up-to-date.